Separate parse logic into a different module
